Hirst Spot Painting Generator/requirements.txt
Pillow>=9.0.0
colorgram.py